refactor: rename home page and controller
