+++ src/main/data/stages.js
